refactor: read record builder code from actual python files
