+++ src/generators/python.ts
@@ -8,6 +8,7 @@ import { Order, PythonGenerator, pythonGenerator } from "blockly/python"
 import * as Blockly from "blockly/core"
 import * as HmcProfile from "../blocks/hmc_profile"
 import * as Util from "./util"
+import builderCode from './record_builder.py'
 
 /**
  * Specialized generator for our code.
@@ -24,50 +25,7 @@ export class RecordMappingGenerator
         this.addReservedWords("math,random,Number")
         Object.assign(this.forBlock, pythonGenerator.forBlock)
         Object.assign(this.forBlock, forBlock)
-        this.definitions_["record-bucket-class"] = `
-class PidRecord:
-    def __init__(self):
-        self._id = ""
-        self._pid = ""
-        self._tuples = set()
-
-    def setPid(self, pid):
-        self._pid = pid
-        return self
-
-    def setId(self, id):
-        self._id = id
-        return self
-
-    def add(self, a: str, b: str | list):
-        if not b or b is None:
-            return self
-        if isinstance(b, list):
-            for item in b:
-                if item is None:
-                    continue
-                self._tuples.add((a, item))
-        else:
-            self._tuples.add((a, b))
-        return self
-
-    def toSimpleJSON(self):
-        result = {"entries": list(self._tuples)}
-        if self._pid and self._pid != "":
-            result["pid"] = self._pid
-        return result
-
-records_graph = []
-`
-        this.definitions_["typed-pid-maker-connections"] = `
-def createSingleRecord(pidrecord):
-    # TODO implement request to a typed PID Maker instance
-    # pseudocode:
-    # if pidrecord.hasPid: update(pidrecord)
-    # else: create(pidrecord)
-    # onError: to be decided
-    return "pid-of-pidrecord"
-`
+        this.definitions_["record-builder-code"] = builderCode
     }
 
     makeAddAttributeChainCall(key: string, value: string): string {
